Moving mapper file per PR suggestion.
