+++ careconnect2025/backend/core/src/main/java/com/careconnect/service/v2/TaskServiceV2.java
@@ -28,6 +28,7 @@ import com.careconnect.model.ScheduledNotification;
 import com.careconnect.model.Task;
 import com.careconnect.repository.PatientRepository;
 import com.careconnect.repository.TaskRepository;
+import com.careconnect.util.TaskMapper;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
