+++ careconnect2025/backend/core/src/main/java/com/careconnect/util/TaskMapper.java
@@ -1,4 +1,4 @@
-package com.careconnect.service.v2;
+package com.careconnect.util;
 
 import java.util.List;
